+++ yash-builtin/src/command.rs
@@ -48,6 +48,7 @@ use enumset::EnumSet;
 use enumset::EnumSetType;
 use yash_env::Env;
 use yash_env::semantics::Field;
+use yash_env::system::concurrency::RunLoop;
 #[cfg(all(doc, unix))]
 use yash_env::system::real::RealSystem;
 use yash_env::system::resource::SetRlimit;
@@ -183,6 +184,7 @@ impl Command {
             + IsExecutableFile
             + Isatty
             + Open
+            + RunLoop
             + SendSignal
             + SetPgid
             + SetRlimit
@@ -225,6 +227,7 @@ where
         + IsExecutableFile
         + Isatty
         + Open
+        + RunLoop
         + SendSignal
         + SetPgid
         + SetRlimit